+++ Client/Client.h
@@ -23,7 +23,7 @@ public:
     static Client *getInstance();
     const std::string& getId() const;
     static const std::string& getConfigPath();
-    void shareFile(const std::string& fileName, const std::string& trackerAddres, const size_t pieceSize);
+    void shareFile(const std::string& fileName, const std::string& trackerAddres, const uint64_t pieceSize);
     void downloadFile(const std::string& torrentPath, const std::string& downloadPath);
 public slots:
     void onFileCreated(TorrentFile*);
